+++ src/Mutation.tsx
@@ -2,12 +2,13 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import ApolloClient, { PureQueryOptions, ApolloError } from 'apollo-client';
 import { DataProxy } from 'apollo-cache';
-const invariant = require('invariant');
+import invariant from 'invariant';
 import { DocumentNode, GraphQLError } from 'graphql';
 const shallowEqual = require('fbjs/lib/shallowEqual');
 
 import { OperationVariables, RefetchQueriesProviderFn } from './types';
 import { parser, DocumentType } from './parser';
+import { getClient } from './component-utils';
 
 export interface MutationResult<TData = Record<string, any>> {
   data?: TData;
@@ -17,7 +18,7 @@ export interface MutationResult<TData = Record<string, any>> {
   client: ApolloClient<Object>;
 }
 export interface MutationContext {
-  client: ApolloClient<Object>;
+  client?: ApolloClient<Object>;
   operations: Map<string, { query: DocumentNode; variables: any }>;
 }
 
@@ -54,6 +55,7 @@ export declare type MutationFn<TData = any, TVariables = OperationVariables> = (
 ) => Promise<void | FetchResult<TData>>;
 
 export interface MutationProps<TData = any, TVariables = OperationVariables> {
+  client?: ApolloClient<Object>;
   mutation: DocumentNode;
   ignoreResults?: boolean;
   optimisticResponse?: Object;
@@ -67,7 +69,6 @@ export interface MutationProps<TData = any, TVariables = OperationVariables> {
   ) => React.ReactNode;
   onCompleted?: (data: TData) => void;
   onError?: (error: ApolloError) => void;
-  client?: ApolloClient<Object>;
   context?: Record<string, any>;
 }
 
@@ -116,17 +117,8 @@ class Mutation<TData = any, TVariables = OperationVariables> extends React.Compo
 
   constructor(props: MutationProps<TData, TVariables>, context: any) {
     super(props, context);
-
-    this.client = props.client || context.client;
-    invariant(
-      !!this.client,
-      'Could not find "client" in the context or props of Mutation. Wrap ' +
-        'the root component in an <ApolloProvider>, or pass an ApolloClient ' +
-        'instance in via props.',
-    );
-
+    this.client = getClient(props, context);
     this.verifyDocumentIsMutation(props.mutation);
-
     this.mostRecentMutationId = 0;
     this.state = initialState;
   }
@@ -143,11 +135,8 @@ class Mutation<TData = any, TVariables = OperationVariables> extends React.Compo
     nextProps: MutationProps<TData, TVariables>,
     nextContext: MutationContext,
   ) {
-    const { client } = nextProps;
-    if (
-      shallowEqual(this.props, nextProps) &&
-      (this.client === client || this.client === nextContext.client)
-    ) {
+    const nextClient = getClient(nextProps, nextContext);
+    if (shallowEqual(this.props, nextProps) && this.client === nextClient) {
       return;
     }
 
@@ -155,8 +144,8 @@ class Mutation<TData = any, TVariables = OperationVariables> extends React.Compo
       this.verifyDocumentIsMutation(nextProps.mutation);
     }
 
-    if (this.client !== client && this.client !== nextContext.client) {
-      this.client = client || nextContext.client;
+    if (this.client !== nextClient) {
+      this.client = nextClient;
       this.setState(initialState);
     }
   }
@@ -178,7 +167,6 @@ class Mutation<TData = any, TVariables = OperationVariables> extends React.Compo
 
   private runMutation = (options: MutationOptions<TVariables> = {}) => {
     this.onMutationStart();
-
     const mutationId = this.generateNewMutationId();
 
     return this.mutate(options)