+++ src/types.ts
@@ -10,6 +10,7 @@ import ApolloClient, {
   PureQueryOptions,
   MutationUpdaterFn,
 } from 'apollo-client';
+import { MutationFn } from './Mutation';
 
 export type OperationVariables = {
   [key: string]: any;
@@ -31,6 +32,8 @@ export interface MutationOpts<TData = any, TGraphQLVariables = OperationVariable
   client?: ApolloClient<any>;
   notifyOnNetworkStatusChange?: boolean;
   context?: Record<string, any>;
+  onCompleted?: (data: TData) => void;
+  onError?: (error: ApolloError) => void;
 }
 
 export interface QueryOpts<TGraphQLVariables = OperationVariables> {
@@ -59,9 +62,11 @@ export interface GraphqlQueryControls<TGraphQLVariables = OperationVariables> {
   updateQuery: (mapFn: (previousQueryResult: any, options: UpdateQueryOptions<any>) => any) => void;
 }
 
-export type MutationFunc<TData = any, TGraphQLVariables = OperationVariables> = (
-  opts?: MutationOpts<TData, TGraphQLVariables>,
-) => Promise<ApolloQueryResult<TData>>;
+// XXX remove in the next breaking semver change (3.0)
+export type MutationFunc<TData = any, TVariables = OperationVariables> = MutationFn<
+  TData,
+  TVariables
+>;
 
 export type DataValue<TData, TGraphQLVariables = OperationVariables> = GraphqlQueryControls<
   TGraphQLVariables
@@ -76,7 +81,7 @@ export interface DataProps<TData, TGraphQLVariables = OperationVariables> {
 
 // export to allow usage individually for simple components
 export interface MutateProps<TData = any, TGraphQLVariables = OperationVariables> {
-  mutate: MutationFunc<TData, TGraphQLVariables>;
+  mutate: MutationFn<TData, TGraphQLVariables>;
 }
 
 export type ChildProps<TProps = {}, TData = {}, TGraphQLVariables = OperationVariables> = TProps &