+++ src/browser.ts
@@ -24,5 +24,4 @@ export * from './withApollo'
 export * from './types';
 
 // XXX remove in the next breaking semver change (3.0)
-const compose = require('lodash/flowRight');
-export { compose };
+export { default as compose } from 'lodash/flowRight';